pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contreact-ollama"
version = "0.1.0"
description = "Replicate 'What Do LLM Agents Do When Left Alone?' research using local Ollama LLMs"
authors = [
    {name = "ContReAct-Ollama Team"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "streamlit>=1.38.0,<2.0.0",
    "tinydb>=4.8.0,<5.0.0",
    "ollama>=0.4.0,<0.5.0",
    "sentence-transformers>=3.0.1,<4.0.0",
    "numpy>=1.26.4",
    "scipy>=1.13.1",
    "plotly>=5.22.0,<6.0.0",
    "pandas>=2.2.2,<3.0.0",
    "pyyaml>=6.0,<7.0",
    "python-telegram-bot>=20.0,<21.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.2,<9.0.0",
    "pytest-playwright>=0.5.0,<1.0.0",
    "black>=24.0.0,<25.0.0",
    "isort>=5.13.0,<6.0.0",
    "mypy>=1.10.0,<2.0.0",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312", "py313"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.setuptools.packages.find]
where = ["."]
include = ["contreact_ollama*"]
exclude = ["tests*", "docs*", "scripts*"]
